Hotfix Gradio return list fix
Gradio dropdown function now returns a list of length 0 rather then an empty string when passed with None. This means it can never call the default constructor if the dropdown is passed into an event. When the hell did they update this. In the documentation it says it will always return either String or Int if using type='index'.
also when setting value="" then it just uses the first element of the list rather then a literal "" like its supposed to.